issues pallet tests passed with set_exchange_rate. update oracle testing (#195)
